refactor(challenge): dropdown/·menu/ 타입버킷 해체 → 사용처(목적)로 흡수
위젯 타입 이름 폴더(dropdown/menu)를 없애고 실사용처로:
- mission/: AbsentContentsDropdown, MenuContentsDropdown, AbsentContentsInfoMenu,
  AbsentMissionSubmitMenu (미션 상세/결석 메뉴가 mission/에서 사용)
- daily-mission/: ContentsDropdown (DailyMissionInfoSection만 사용)
mission/ 내부는 형제라 ./ 상대경로, daily-mission은 경로 remap. tsc 0 / lint 통과.
