+++ apps/web/src/domain/challenge/my-challenge/mission/DoneMissionDetailMenu.tsx
@@ -8,7 +8,7 @@ import {
 import { BONUS_MISSION_TH } from '@/utils/constants';
 import { useParams } from 'next/navigation';
 import OtVideo from '@/domain/challenge/my-challenge/ot-mission/OtVideo';
-import MenuContentsDropdown from '../dropdown/MenuContentsDropdown';
+import MenuContentsDropdown from './MenuContentsDropdown';
 import ParsedCommentBox from '../ui/ParsedCommentBox';
 interface Props {
   missionDetail: UserChallengeMissionDetail;
